refactor: extract TrackerExportRetryContext.swift from TrackerExportRetryTypes.swift
Splits the `TrackerExportRetryContext` Sendable value (per-provider
naming for the shared retry runner) out of TrackerExportRetryTypes.swift.

Byte-identical move of lines 23-37 (includes the load-bearing
provider-parity doc comment). Both files import Foundation.

TrackerExportRetryTypes.swift: 37 → 22 lines. Tests: 667.
